EBCD-13341 . a bit of cleanup
